+++ martin-mbtiles/src/lib.rs
@@ -15,8 +15,8 @@ pub use pool::MbtilesPool;
 mod copier;
 pub use copier::{CopyDuplicateMode, MbtilesCopier};
 
-mod applier;
-pub use applier::apply_diff;
+mod patcher;
+pub use patcher::apply_diff;
 
 mod queries;
 pub use queries::{
